refactor dagger out of cli
