+++ atlasdb-cli/build.gradle
@@ -14,8 +14,6 @@ dependencies {
   runtime project(':atlasdb-rocksdb')
 
   testCompile project(':atlasdb-rocksdb')
-
-  processor 'com.google.dagger:dagger-compiler:' + libVersions.dagger
 }
 
 configurations.matching({ it.name in ['compile', 'runtime'] }).all {
