+++ atlasdb-cli/src/test/java/com/palantir/atlasdb/cli/command/TestSweepCommand.java
@@ -28,9 +28,6 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.io.BaseEncoding;
 import com.palantir.atlasdb.cli.runner.InMemoryTestRunner;
 import com.palantir.atlasdb.cli.runner.SingleBackendCliTestRunner;
-import com.palantir.atlasdb.cli.services.DaggerTestAtlasDbServices;
-import com.palantir.atlasdb.cli.services.TestAtlasDbServices;
-import com.palantir.atlasdb.cli.services.TestSweeperModule;
 import com.palantir.atlasdb.keyvalue.api.Cell;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.api.Namespace;
@@ -39,6 +36,9 @@ import com.palantir.atlasdb.keyvalue.api.Value;
 import com.palantir.atlasdb.protos.generated.TableMetadataPersistence;
 import com.palantir.atlasdb.services.AtlasDbServicesFactory;
 import com.palantir.atlasdb.services.ServicesConfigModule;
+import com.palantir.atlasdb.services.test.DaggerTestAtlasDbServices;
+import com.palantir.atlasdb.services.test.TestAtlasDbServices;
+import com.palantir.atlasdb.services.test.TestSweeperModule;
 import com.palantir.atlasdb.table.description.TableDefinition;
 import com.palantir.atlasdb.table.description.ValueType;
 import com.palantir.atlasdb.transaction.api.ConflictHandler;
